formをコンポーネント化
