+++ apps/frontend/src/app/biz/(auth)/login/page.tsx
@@ -1,51 +1,9 @@
-"use client";
-
-import Link from "next/link";
+import { BizLoginForm } from '@/app/components/features/auth/BizLoginForm';
 
 export default function BusinessLogIn() {
   return (
-    <div className="container mx-auto max-w-md px-4 py-8">
-      <h2 className="text-2xl font-bold text-center mb-6">ログイン</h2>
-      <form className="space-y-6">
-        <div>
-          <label className="block text-sm font-medium text-gray-700 mb-1">
-            メールアドレス
-          </label>
-          <input
-            id="email"
-            type="email"
-            className="w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-blue-500 focus:border-blue-500"
-            placeholder="abc@test.com"
-          />
-        </div>
-        <div>
-          <label className="block text-sm font-medium text-gray-700 mb-1">
-            パスワード
-          </label>
-          <input
-            id="password"
-            type="password"
-            className="w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-blue-500 focus:border-blue-500"
-            placeholder="●●●●●●●●"
-          />
-        </div>
-        <div>
-          <button
-            type="submit"
-            className="w-full py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-blue-600 hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500"
-          >
-            ログインする
-          </button>
-        </div>
-      </form>
-      <div className="mt-6 text-center">
-        <p className="text-sm text-gray-600">アカウントがない場合</p>
-        <Link href="/business/register">
-          <button className="mt-2 w-full py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-green-600 hover:bg-green-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-green-500">
-            新規登録する
-          </button>
-        </Link>
-      </div>
+    <div className="container mx-auto max-w-md px-6 py-10">
+      <BizLoginForm />
     </div>
   );
 }
